% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bulktest}
\alias{bulktest}
\title{bulktest}
\format{
An object of class \code{data.frame} with 11953 rows and 10 columns.
}
\usage{
bulktest
}
\description{
From part of GSE14520
}
\examples{
  head(bulktest)
}
\keyword{datasets}
