% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_set}
\alias{gene_set}
\title{gene_set}
\format{
An object of class \code{list} of length 2.
}
\usage{
gene_set
}
\description{
A list of C1 and C2 gene sets
}
\examples{
  head(gene_set)
}
\keyword{datasets}
