% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KNN}
\alias{KNN}
\title{KNN}
\format{
An object of class \code{train} (inherits from \code{train.formula}) of length 24.
}
\usage{
KNN
}
\description{
K nearest neighbors model
}
\examples{
  str(KNN)
}
\keyword{datasets}
