% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pair_select}
\alias{pair_select}
\title{pair_select}
\format{
An object of class \code{data.frame} with 97 rows and 3 columns.
}
\usage{
pair_select
}
\description{
Gene pairs for machine learning models
}
\examples{
  head(pair_select)
}
\keyword{datasets}
