% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_29pairs.R
\name{predict_29pairs}
\alias{predict_29pairs}
\title{Predicted immuno-prognostic subtypes in bulk transcriptomic data.}
\usage{
predict_29pairs(data, Pairs, genetype = "Symbol")
}
\arguments{
\item{data}{Composed of gene symbol or id and exp.}

\item{Pairs}{29 gene pairs}

\item{genetype}{Symbol or ID}
}
\value{

}
\description{
Predicted immuno-prognostic subtypes in bulk transcriptomic data.
}
\examples{
test <- predict_29pairs(bulktest,Pairs,genetype = "Symbol")

}
