% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_scrna.R
\name{predict_scrna}
\alias{predict_scrna}
\title{Predicted immuno-prognostic subtypes in single-cell data.}
\usage{
predict_scrna(data, split.by = "orig.ident", method = "All")
}
\arguments{
\item{data}{Seurat object.}

\item{split.by}{Sample class}

\item{method}{knn, nb, mean, All Default: All (including three models)}
}
\value{

}
\description{
Predicted immuno-prognostic subtypes in single-cell data.
}
\examples{
test <- predict_scrna(SingleTest,split.by = "orig.ident", method = "All")
}
