% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RF}
\alias{RF}
\title{RF}
\format{
An object of class \code{train} (inherits from \code{train.formula}) of length 24.
}
\usage{
RF
}
\description{
Random forest model
}
\examples{
  str(RF)
}
\keyword{datasets}
