% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SingleTest}
\alias{SingleTest}
\title{SingleTest}
\format{
An object of class \code{Seurat} with 25712 rows and 200 columns.
}
\usage{
SingleTest
}
\description{
From part of GSE149614
}
\examples{
  head(SingleTest)
}
\keyword{datasets}
