% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_scrna.R
\name{predict_scrna}
\alias{predict_scrna}
\title{Predicted immuno-prognostic subtypes in single-cell data.}
\usage{
predict_scrna(data, gene_set, split.by = "orig.ident", method = "All")
}
\arguments{
\item{data}{Seurat object. Please use FindVariableFeatures function from Seurat to identify highly variable features !!!}

\item{gene_set}{A list of C1 and C2 gene sets}

\item{split.by}{Sample class}

\item{method}{ssGSEA, RF, KNN. Default: All (including three models)}
}
\value{

}
\description{
Predicted immuno-prognostic subtypes in single-cell data.
}
\examples{

}
